refactor(u7agent): rename component file to match new component name
