+++ projects/u7agent/src/features/chat/chat.tsx
@@ -7,7 +7,7 @@ import { ChatInput } from './chat-input'
 import { ChatMessage } from './chat-message'
 import { DebugMessagesOverlay } from './debug-messages-overlay'
 import { JumpToLatestButton } from './jump-to-latest-button'
-import { ModelRunSummary } from './token-usage-summary'
+import { ModelRunSummary } from './model-run-summary'
 import { useAgentStreamChat } from './use-agent-stream-chat'
 import { useChatScroll } from './use-chat-scroll'
 import { env } from '@/env'
